chore(startup): clean startup file
