+++ Micro.KeyStore.Api/StartupExtensions/Logging.cs
@@ -1,0 +1,24 @@
+using System;
+using Micro.KeyStore.Api.Configs;
+using Microsoft.Extensions.Hosting;
+using Microsoft.Extensions.Logging;
+using Microsoft.Extensions.Logging.Slack;
+
+namespace Micro.KeyStore.Api.StartupExtensions
+{
+    public static class Logging
+    {
+        public static void ConfigureLoggerWithSlack(this ILoggerFactory loggerFactory, SlackLoggingConfig slackConfig, IHostEnvironment env)
+        {
+            if (string.IsNullOrEmpty(slackConfig.WebhookUrl))
+            {
+                return;
+            }
+            loggerFactory.AddSlack(new SlackConfiguration
+            {
+                MinLevel = slackConfig.MinLogLevel,
+                WebhookUrl = new Uri(slackConfig.WebhookUrl)
+            }, env.ApplicationName, env.EnvironmentName);
+        }
+    }
+}
