+++ Micro.KeyStore.Api/StartupExtensions/Workers.cs
@@ -1,0 +1,13 @@
+using Micro.KeyStore.Api.Workers;
+using Microsoft.Extensions.DependencyInjection;
+
+namespace Micro.KeyStore.Api.StartupExtensions
+{
+    public static class Workers
+    {
+        public static void RegisterWorkers(this IServiceCollection services)
+        {
+            services.AddHostedService<CleanupKeysWorker>();
+        }
+    }
+}
